[PACKAGES-MIGRATION] move git-metadata to base
