+++ packages/base/package.json
@@ -39,6 +39,7 @@
   },
   "dependencies": {
     "@types/datadog-metrics": "0.6.1",
+    "@types/semver": "^7.7.1",
     "async-retry": "1.3.1",
     "axios": "^1.11.0",
     "chalk": "3.0.0",
