+++ packages/base/src/commands/git-metadata/__tests__/fixtures/add_delete.diff
