+++ packages/base/src/commands/git-metadata/__tests__/fixtures/modify_single_file.diff
