+++ packages/base/src/commands/git-metadata/git.ts
@@ -1,8 +1,9 @@
 import {URL} from 'url'
 
-import {gitRemote} from '@datadog/datadog-ci-base/helpers/git/get-git-data'
 import * as simpleGit from 'simple-git'
 import {GitConfigScope} from 'simple-git'
+
+import {gitRemote} from '../../helpers/git/get-git-data'
 
 import {CommitInfo} from './interfaces'
