+++ packages/base/src/commands/git-metadata/gitdb.ts
@@ -5,14 +5,15 @@ import os from 'os'
 
 import type {AxiosResponse} from 'axios'
 
-import {getDefaultRemoteName, gitRemote as getRepoURL} from '@datadog/datadog-ci-base/helpers/git/get-git-data'
-import {RequestBuilder} from '@datadog/datadog-ci-base/helpers/interfaces'
-import {Logger} from '@datadog/datadog-ci-base/helpers/logger'
-import {retryRequest} from '@datadog/datadog-ci-base/helpers/retry'
 import FormData from 'form-data'
 import {lte} from 'semver'
 import * as simpleGit from 'simple-git'
 import upath from 'upath'
+
+import {getDefaultRemoteName, gitRemote as getRepoURL} from '@datadog/datadog-ci-base/helpers/git/get-git-data'
+import {RequestBuilder} from '@datadog/datadog-ci-base/helpers/interfaces'
+import {Logger} from '@datadog/datadog-ci-base/helpers/logger'
+import {retryRequest} from '@datadog/datadog-ci-base/helpers/retry'
 
 const API_TIMEOUT = 15000
