+++ packages/base/src/commands/git-metadata/interfaces.ts
