+++ packages/base/src/commands/git-metadata/renderer.ts
@@ -1,5 +1,6 @@
-import {ICONS} from '@datadog/datadog-ci-base/helpers/formatting'
 import chalk from 'chalk'
+
+import {ICONS} from '@datadog/datadog-ci-base/helpers/formatting'
 
 import {CommitInfo} from './interfaces'
