+++ packages/datadog-ci/src/commands/flutter-symbols/upload.ts
@@ -1,5 +1,6 @@
 import fs from 'fs'
 
+import {newSimpleGit} from '@datadog/datadog-ci-base/commands/git-metadata/git'
 import {FIPS_ENV_VAR, FIPS_IGNORE_ERROR_ENV_VAR} from '@datadog/datadog-ci-base/constants'
 import {newApiKeyValidator} from '@datadog/datadog-ci-base/helpers/apikey'
 import {doWithMaxConcurrency} from '@datadog/datadog-ci-base/helpers/concurrency'
@@ -24,7 +25,6 @@ import semver from 'semver'
 import {cliVersion} from '../../version'
 
 import * as dsyms from '../dsyms/upload'
-import {newSimpleGit} from '../git-metadata/git'
 import * as sourcemaps from '../sourcemaps/upload'
 
 import {getArchInfoFromFilename, getFlutterRequestBuilder, uploadMultipartHelper} from './helpers'
