+++ packages/datadog-ci/src/index.ts
@@ -1,4 +1,4 @@
-export * as gitMetadata from './commands/git-metadata'
+export * as gitMetadata from '../../base/src/commands/git-metadata'
 export * as synthetics from '@datadog/datadog-ci-plugin-synthetics'
 export * as utils from '@datadog/datadog-ci-base/helpers/utils'
 export {cliVersion as version} from './version'
